--- src/database/Identifiable.java
@@ -1,10 +1,0 @@
-package database;
-
-/**
- * Interface that represents the any record that can be uniquely identified in the database.
- */
-public interface Identifiable {
-    
-    // Returns the unique key associated with the record
-    public abstract int getId();
-}
